Pulling custom query handling and parsing out to a separate experiment file
